Move head and tail of of cint class.
Remove head and tail members from cint and make them free standing
metafunctions. The goal is to make all operations on cints not require
the `cint` type to be complete. The type would only be completed in
cint.h as the last step and would be empty except for cast operators.