Allow conversions of cint to integral types and enums.
This change introduces a new metafunction, `_cint::convert`, which
defines a member constant `value` with the same value as the provided
cint. If the cint doesn't fit into the target type, the value is
truncated and the `implicit` member value is set to `false`.

The definition of `cint` then defines both implicit and explicit
versions of the conversion operator based on the new metafunction.